bibtrim - a script for trimming verbose .bib files.
I find that, to get all of the information I need in a .bib file, I have to export everything that Papers knows how to export. The downside of this approach is that the resulting files are a real pain to merge.

This script trims out superfluous information from Papers-generated BibTeX files and also makes a couple of other tweaks to work around bizarre behaviour.